requirements.txt
numpy
opencv-python
pyocr
pillow
fastapi
python-multipart